+++ gems/aws-sdk-core/lib/aws-sdk-core/shared_config.rb
@@ -212,6 +212,7 @@ module Aws
       :ec2_metadata_service_endpoint,
       :ec2_metadata_service_endpoint_mode,
       :ec2_metadata_v1_disabled,
+      :disable_host_prefix_injection,
       :max_attempts,
       :retry_mode,
       :adaptive_retry_wait_to_fill,